Move release specific searches to the release package
