+++ cli/pkg/gbm/search.go
@@ -1,43 +1,12 @@
 package gbm
 
 import (
-	"fmt"
 	"strings"
 
 	"github.com/wordpress-mobile/gbm-cli/pkg/console"
 	"github.com/wordpress-mobile/gbm-cli/pkg/gh"
-	"github.com/wordpress-mobile/gbm-cli/pkg/repo"
 )
 
-func FindGbmReleasePr(version string) (gh.PullRequest, error) {
-	label := fmt.Sprintf("label:%s", GbmReleasePrLabel)
-	title := fmt.Sprintf("%s in:title", version)
-
-	filter := gh.BuildRepoFilter(repo.GutenbergMobileRepo, "is:pr", "is:open", label, title)
-	pr, err := gh.SearchPr(filter)
-	if err != nil {
-		return gh.PullRequest{}, err
-	}
-	pr.ReleaseVersion = version
-	return pr, nil
-}
-
-func FindAndroidReleasePr(version string) (gh.PullRequest, error) {
-	label := fmt.Sprintf("label:%s", IntegrationPrLabel)
-	title := fmt.Sprintf("v%s in:title", version)
-
-	filter := gh.BuildRepoFilter(repo.WordPressAndroidRepo, "is:open", "is:pr", label, title)
-
-	return gh.SearchPr(filter)
-}
-
-func FindIosReleasePr(version string) (gh.PullRequest, error) {
-	label := fmt.Sprintf("label:%s", IntegrationPrLabel)
-	title := fmt.Sprintf("v%s in:title", version)
-
-	filter := gh.BuildRepoFilter(repo.WordPressIosRepo, "is:open", "is:pr", label, title)
-	return gh.SearchPr(filter)
-}
 func FindGbmSyncedPrs(gbmPr gh.PullRequest, filters []gh.RepoFilter) ([]gh.SearchResult, error) {
 	var synced []gh.SearchResult
 	prChan := make(chan gh.SearchResult)
@@ -71,8 +40,4 @@ func FindGbmSyncedPrs(gbmPr gh.PullRequest, filters []gh.RepoFilter) ([]gh.Searc
 	}
 
 	return synced, nil
-}
-
-func GetGbmRelease(version string) (gh.Release, error) {
-	return gh.GetReleaseByTag(repo.GutenbergMobileRepo, "v"+version)
 }
